fix : 수익률 오류 수정
리팩토링에서 파라미터 값을 잘못 넣어줘서 오류났던 수익률 계산을 수정했다.

+++ src/main/java/lotto/ui/OutputHandler.java
@@ -32,8 +32,8 @@ public class OutputHandler {
         }
     }
 
-    public void printEarningsRate(PurchaseAmount amount, int money){
-        float rate = amount.calculateTotalProfitRate(money) ; // 수익률을 퍼센트로 변경
+    public void printEarningsRate(PurchaseAmount amount, long money){
+        float rate = amount.calculateTotalProfitRate(money) ;
         if (rate % 1.0 == 0) {
             System.out.printf("총 수익률은 %.0f%% 입니다.\n", rate);
         } else {
